### 2024-11-21 #7317 @elbeejay
@Scotsman1973 I am going to begin contacting potential reviewers to ask if they are willing and able to help review your submission.

Reading the submitted paper I had a few small suggestions:

- Line 17: consider pluralizing "application"
- Capitalize references to figures in the text, e.g., (Figure 1) as opposed to (figure 1)
- Line 54: consider adding "the" before "data's"
