### 2024-11-21 #7317 @elbeejay
@jtagusari would you be willing to review this JOSS submission, "archaeoModelling: a QGIS plugin using the MaxEnt algorithm"? Your work developing QGIS plugin applications, as well as your quantitative background more generally should make you a highly qualified reviewer for this submission.

I know you published a QGIS plugin on H-RISK and NoiseModelling with JOSS, so I expect you're familiar with our open approach to reviewing. Please let us know if you are able and willing to review.

If you cannot provide your expertise as a reviewer, please take a minute and let us know if you have any colleagues, peers, or trainees who might be suitable reviewers.

Thanks,
Jay
