### 2024-11-21 #7317 @elbeejay
@MichaelHoltonPrice would you be willing to review this JOSS submission, "archaeoModelling: a QGIS plugin using the MaxEnt algorithm"? Given your background at this nexus of geospatial science and statistics, I think you'd be a strong reviewer.

From what I can tell you've reviewed for JOSS in the past, so I expect you're familiar with our open approach to reviewing. Please let us know if you are able and willing to review, at this time I have been trying to ask reviewers to complete their first reviews within 6 weeks of us starting the process.

If you cannot provide your expertise as a reviewer, please take a minute and let us know if you have any colleagues, peers, or trainees who might be suitable reviewers.

Thanks,
Jay
