+++ paper/paper.md
@@ -14,7 +14,7 @@ authors:
   - name: Andrea Jalandoni
     orcid: 0000-0002-4821-7183
     corresponding: false
-    affiliation: "1, 2"
+    affiliations: "1, 2"
 affiliations:
  - name: Australian Research Centre for Human Evolution, Griffith University, Australia
    index: 1
@@ -39,9 +39,9 @@ Often APM falls short of its potential, because it can be difficult to explain h
 
 # Functionality
 
-```archaeoModelling``` was built within the ```QGIS``` framework (figure 1) using Python and the ```Qt Designer``` application, along with the [Plugin Builder](https://github.com/g-sherman/Qgis-Plugin-Builder) and [PluginReloader](https://github.com/borysiasty/plugin_reloader) tools, and its Graphical User Interface (GUI) wraps ```elapid``` [@Anderson:2023], a Python package designed for species distribution modelling using MaxEnt.  Additionally, the geospatial processing was developed specifically for archaeological predictive modelling.  Inputs include the ‘activity area’, where an excavation or survey will occur, and the ‘geographic area’, the area that geospatial data, such as elevation and landform, refers to (figure 2).  The activity area can be the same as the geographic area or within its boundaries, but no part can be outside the boundaries.  For simplicity, a landform vector file (polygon or multipolygon) can be used as the geographic area.
+```archaeoModelling``` was built within the ```QGIS``` framework (Figure 1) using Python and the ```Qt Designer``` application, along with the [Plugin Builder](https://github.com/g-sherman/Qgis-Plugin-Builder) and [PluginReloader](https://github.com/borysiasty/plugin_reloader) tools, and its Graphical User Interface (GUI) wraps ```elapid``` [@Anderson:2023], a Python package designed for species distribution modelling using MaxEnt.  Additionally, the geospatial processing was developed specifically for archaeological predictive modelling.  Inputs include the ‘activity area’, where an excavation or survey will occur, and the ‘geographic area’, the area that geospatial data, such as elevation and landform, refers to (Figure 2).  The activity area can be the same as the geographic area or within its boundaries, but no part can be outside the boundaries.  For simplicity, a landform vector file (polygon or multipolygon) can be used as the geographic area.
 
-When processing, ```archaeoModelling``` accesses data’s attribute table (figure 3, Geospatial inputs), so, codes, terminology and any mistakes are the responsibility of the body producing or altering data.  Interpretation of landforms (e.g. soil) can be quantified, users giving unique features (e.g. sand or peat) a 1 – 9 weight based on site formation and destruction [@Boemke:2022].  The GUI’s default settings rank everything similar to the activity area higher than that which is different.  For example, if an activity area’s slope is 5 – 15 degrees then everywhere in the geographic area with a slope between 5 – 15 degrees will be ranked higher than elsewhere.  However, weights created by default settings can be changed within the GUI.  The results of processing are rasterized, then saved in a nominated directory (figure 3, folder for model). Models are created from all rasters in the directory, limited only by computational power.
[…]
