### 2024-11-25 #7317 @elbeejay
@skiptoniam - would you consider reviewing the submission "archaeoModelling: a QGIS plugin using the MaxEnt algorithm" for JOSS? This is a bit outside of my realm of expertise, so I apologize if it isn't a great fit, but the spatial nature of the work reminded me of your `ppmData` package. Please let us know if you feel you can provide a good review, and if you are able and willing to serve as a reviewer.

If not, but you have some suggestions of good potential reviewers for this submission, please let me know, thanks!
