### 2024-12-04 #7317 @btupper
Thanks for asking, but I, too, will decline at this time.

Ben

On Tue, Dec 3, 2024 at 5:32 PM J. Hariharan ***@***.***>
wrote:

> @goldingn <https://github.com/goldingn> and @btupper
> <https://github.com/btupper>, would you be interested in reviewing this
> submission to JOSS titled "archaeoModelling: a QGIS plugin using the MaxEnt
> algorithm"? JOSS is an open-source, diamond access software journal. I saw
> that you've both contributed to the R library maxnet which I believe
> applies the same MaxEnt algorithm here for species modeling. Given your
> interest in this particular algorithm/method, as well as your more general
> expertise in computing and geospatial science, I thought you might be good
> reviewers for this submission.
>
> At JOSS we do open checklist-driven reviews; peer-review criteria can be
> viewed here <https://joss.readthedocs.io/en/latest/review_criteria.html>.
> This issue is a "pre-review issue" which we use to find peer-reviewers.
> Once 3 reviewers are found, we will officially start the review in a
> dedicated GitHub issue. At present we are asking reviewers to complete
> reviews in 6 weeks, although this can be extended if needed. If you are not
> able to review but can recommend someone else, please mention them here (in
> this case please mention their GitHub handle without the "@" symbol).
>
> If you are interested, please take a look at the journal's conflict of
> interest policy
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy>
> to ensure you do not have a conflict before agreeing to review this
> submission.
>
> Please do not feel any pressure to accept this review request if you do
> not have the time or do not feel comfortable reviewing this software
> package, we appreciate and respect our peer-reviewers' time. If you cannot
> serve as a reviewer at this time but have a peer, collaborator, student, or
> colleague who might be available and would be a good fit for this
> submission, please let me know!
>
> Thanks,
> Jay
>
