### 2024-12-04 #7317 @skiptoniam
Sorry I only just saw this email. I can't review it right now, it's a bit
crazy towards the end of the year and I probably don't have the QGIS skills
to know what a good plug in looks like.

Cheers,

Skip.

On Tue, Nov 26, 2024 at 12:57 AM J. Hariharan ***@***.***>
wrote:

> @skiptoniam <https://github.com/skiptoniam> - would you consider
> reviewing the submission "archaeoModelling: a QGIS plugin using the MaxEnt
> algorithm" for JOSS? This is a bit outside of my realm of expertise, so I
> apologize if it isn't a great fit, but the spatial nature of the work
> reminded me of your ppmData package. Please let us know if you feel you
> can provide a good review, and if you are able and willing to serve as a
> reviewer.
>
> If not, but you have some suggestions of good potential reviewers for this
> submission, please let me know, thanks!
>
