### 2024-12-12 #7587 @editorialbot
[REVIEW]: archaeoModelling: a QGIS plugin using the MaxEnt algorithm
**Submitting author:** @Scotsman1973 (<a href="http://orcid.org/0009-0001-5964-2652">Andrew Alexander Prentice</a>)
**Repository:** https://github.com/Scotsman1973/archaeo_modelling
**Branch with paper.md** (empty if default branch):
**Version:** v1.1
