### 2024-12-12 #7587 @elbeejay
@jtagusari & @goldingn, reviewer instructions should be in the first comment of this issue, but please do not hesistate to reach out to me if you have any questions or concerns about the process! Like it states up top, we hope to have reviews finished within 6 weeks; in this case, if we could target the end of January that'd be fantastic. Since JOSS has an open review process, feel free to review and make comments in a piece-wise fashion as you have time, or you can review the submission all at once, whatever is most convenient for you. Thanks again for your help!
