### 2025-01-06 #7587 @elbeejay
@jtagusari & @goldingn, just checking in and hopefully providing a friendly reminder about this review for JOSS. Please take a look and put it on your calendars or to-do lists when you get a chance. If you can, let us know roughly when you'll have the time to conduct your review so that @Scotsman1973 and I are aware. Thanks!
