### 2025-01-21 #7587 @goldingn
Hi @Scotsman1973!

Can you please confirm my understanding that this software is only intended to work on Windows machines?

The installation instructions require OSGeo4W (windows-only software) and when I tried to work around that on my mac, installation failed on trying to load the win32api python package ([called here](https://github.com/Scotsman1973/archaeo_modelling/blob/2f8f627a464dc8ce7381bb3a2f60e106bdedb4fa/archaeo_modelling_dialog.py#L34)), which implies it's hard-coded for windows. Please correct me if that's wrong though.

I think the paper is likely still within scope for JOSS even if it's Windows-only. Though the documentation and the paper should state this, if so.

It might also take me some time to review installation and functionality, as I'll need to borrow a windows laptop first.
