### 2025-01-21 Scotsman1973/archaeo_modelling#2 @goldingn
Paper affiliations incorrectly formatted
The paper is failing to compile [in the JOSS review](https://github.com/openjournals/joss-reviews/issues/7587#issuecomment-2538852290) due to one of the affiliations being 'missing'.

I think that's because [this line](https://github.com/Scotsman1973/archaeo_modelling/blob/main/paper/paper.md?plain=1#L17) is formatted with quotes, but shouldn't be, [per the documentation](https://joss.readthedocs.io/en/latest/paper.html#affiliations)
