### 2025-01-21 Scotsman1973/archaeo_modelling#2 @Scotsman1973
Quotes removed
