### 2025-01-21 #7587 @Scotsman1973
Hi Nick,

Thank you for pointing the glitch out.

I do not have a Mac, but will get hold of one for testing and remove the windows specific code

Regards, Andrew
