### 2025-01-21 Scotsman1973/archaeo_modelling#3 @goldingn
Add to documentation
Re. the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7587): to pass the [reviewer checklist](https://github.com/openjournals/joss-reviews/issues/7587#issuecomment-2603814413) will require a lot more information in the [package documentation](https://github.com/Scotsman1973/archaeo_modelling/blob/main/README.md).

Specifically:

- **A statement of need**: there is no information in the documentation saying what the software is for, and who it's for

- **Installation instructions**: this could be a little more verbose and easy to follow. E.g. you could link to the appropriate zenodo zip file download. It would be helpful to spell out more clearly the steps required in unzipping the folder. Ie. "leave 'archaeoModelling.zip' as zip file" refers to another zipfile inside the zipfile just uncompressed (which is named something similar), and "Choose the file" presumably refers to the zipfule 'archaeoModelling.zip' inside the previously uncompressed zipfile. Also "ignore warning" is a little unclear, you should probably let the user know which warning to ignore and why it's safe to do so.

- **Example usage**: There is currently no example usage in the documentation. This is necessary for evaluating the package functionality, in addition to being a requirement for publication.

- **Functionality documentation**: As far as I can tell, there is no information on how to navigate the various pages and boxes in this GUI. Some detailed explanation is needed. It's possible that's provded in the GUI itself, though I haven't yet managed to install it and check.

- **Automated tests**: In addition to unit/integration tests in the software, there should be some information in the documentation for users to ensure they have a working installation. I can't see anything on this yet.

- **Community guidelines**: As far as I can tell, there are no guidelines for contributors and users
