### 2025-01-21 #7587 @goldingn
Thanks Andrew! I'll wait for those updates before assessing installation etc. then
