### 2025-01-21 Scotsman1973/archaeo_modelling#2 @goldingn
paper compiling fine now!
