### 2025-01-21 #7587 @goldingn
Thanks for quick reply and fix to the paper doc @Scotsman1973. I think I've done all I can for now. I look forward to reviewing the software installation and functionality, and the revised documentation, whenever they are ready.
