### 2025-02-05 #7587 @elbeejay
Hello all, checking in here to see how things are going.

@Scotsman1973 are you still in the process of obtaining a Mac and removing the windows specific code so @goldingn can finish their review?

@jtagusari where do you stand on your review? Do you need anything from myself or @Scotsman1973?

Thanks all!
