### 2025-02-05 #7587 @Scotsman1973
Hi,

Yes, windows specific code is being removed, and I'm improving documentation.

Regards, @Scotsman1973
