### 2025-02-06 #7587 @jtagusari
Hello.

I have commented issues (https://github.com/Scotsman1973/archaeo_modelling/issues) and am waiting revision of the paper and the instruction of the software including a tutorial, to check the functionalities.

Thanks,
