### 2025-02-25 #7587 @elbeejay
@Scotsman1973 it looks like things have gotten a bit stagnant both here and in the issues opened up in the `archaeoModelling` repository. Do you have an update for us or an estimate as to when you will be able to address the existing reviewer comments?

Thanks!
